+++ qutip/solver/heom/bofin_baths.py
