+++ qutip/tests/solver/heom/test_bofin_baths.py
@@ -7,7 +7,7 @@ import pytest
 
 from qutip import spre, spost, sigmax, sigmaz
 from qutip.core import data as _data
-from qutip.solve.nonmarkov.bofin_baths import (
+from qutip.solver.heom.bofin_baths import (
     BathExponent,
     Bath,
     BosonicBath,
